==> notebooks/2_Parse_Data.ipynb <==
### Notebook 2 - Parse RaycloudTools Output

This notebook extracts the useful metrics from RCT outputs and saves them to a CSV. It uses several functions for parsing data that are defined in helper_functions.py. The ouputs from this notebook can be used for further analysis in subsequent notebooks and also within other software like R or QGIS.

### Install the required Python packages.
    
Once this has been run once, you can comment out the code using the '#' character at the start of the line.

! pip install pandas matplotlib numpy shapely geopandas scipy

# Suppress depreciation warnings
import warnings
warnings.filterwarnings("ignore", category=DeprecationWarning)

# Import the required modules
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import helper_functions

### Read tree level data from treefile generated from rayextract trees

tree_df = helper_functions.attributes_tree('../data/Mak1_clipped_raycloud_inside_trees.txt')
tree_df.head()

### Read in tree level data from treefile generated using treeinfo. This data file will have the '_trees_info.txt' suffix.

tree_info_df = helper_functions.treeinfo_attributes_tree('../data/Mak1_clipped_raycloud_inside_trees_info.txt')
tree_info_df.head()

### Join the two tree level treefile dataframes

# Join the tree_info_df and tree_df dataframes on the 'tree_id' column
tree_merged_df = tree_info_df.merge(tree_df , on='tree_id')

# Select the columns to be included in the final dataframe
tree_merged_df = tree_merged_df[['x', 'y', 'z','tree_id', 'DBH', 'crown_radius', 'height']]

# Display the first few rows of the final dataframe
tree_merged_df.head()

### Read in the branch segment level data from treefile generated from rayextract trees

branch_segments_df = helper_functions.treeinfo_attributes_segment('../data/Mak1_clipped_raycloud_inside_trees_info.txt')
branch_segments_df.head()

### Calculate total tree volumes by summing the branch segment volumes by 'tree_id'

# Group by 'tree_id' and calculate the sum of 'volume' for each group
total_volume_per_tree = branch_segments_df.groupby('tree_id')['volume'].sum()
total_volume_per_tree

### Join dataframes on the 'tree_id' column

# Join the tree_info_df and tree_df dataframes on the 'tree_id' column
merged_df_final = tree_merged_df.merge(total_volume_per_tree , on='tree_id')
merged_df_final

### Finally save the data to a csv

# Save to csv
merged_df_final.to_csv('../data/rct_tree_data.csv', index=False)
